Move sorting keys to separate files
